+++ HWDescription/CbcRegItem.h
@@ -1,7 +1,7 @@
 #ifndef _CbcRegItem_h__
 #define _CbcRegItem_h__
 
-#include "Rtypes.h"
+#include <cstdint>
 
 namespace Ph2_HwDescription{
 
@@ -11,16 +11,16 @@ namespace Ph2_HwDescription{
 		/* Keep the Name?  Lorenzo: I'm not sure it's necessary to keep the Name here because the variable structure will be
 		already associated with the key of the CbcRegMap wich is already the Name */
 
-		uint32_t fPage;
-		uint32_t fAddress;
-		uint32_t fDefValue;
-		uint32_t fValue;
+		uint8_t fPage;
+		uint8_t fAddress;
+		uint8_t fDefValue;
+		uint8_t fValue;
 
 	};
 
 	// Functor to compare CbcRegItems
-	// Do we need this? If we keep the name it might be uint32_teresting to use a set instead of std::map and do a find?
-	//Lorenzo: i think it's more uint32_teresting to keep std::map, it has already a method find
+	// Do we need this? If we keep the name it might be uint8_teresting to use a set instead of std::map and do a find?
+	//Lorenzo: i think it's more uint8_teresting to keep std::map, it has already a method find
 	struct RegItemComparer {
 
 	    bool operator() (const CbcRegItem &ri1, const CbcRegItem &ri2) const {
